Huge commit. Sorry guys.
The biggest change is a rearrangement of the import structure. Using 'from'
imports in the code was just a bad idea. As was having the Network class reside
in a module called Network.py. So I moved KeyedList.py, Model.py, Network.py,
and Trajectory.py into *_mod.py. Much cleaner.