Removed function substitution code.
In the past, we evaluated expressions by 'exploding' the function definitions
until they reached core functions. This textual manipulation was slow.

Now, we create a 'namespace' dictionary, that contains lambdas of all the
functions we use. We can eval and exec code in this namespace to get access
to those functions. Thus we don't need to 'explode' the code. The result
is cleaner, and runs faster.