Clean-up of _make_res_function.
Also creates two lists in make_res_function
self._residual_terms is a list of what is in each residual, useful for when we want to build derivative functions.
self._dynamic_var_algebraic is a list indicating whether or not dynamic variables are algebraic. This is used when ddaskr calculates initial conditions for us.