Added 'buffer' attribute to events.
If buffer > 0, then after that event executes, integration proceeds for a time buffer with no event detection in place. This particularly helps with events that depend on the time derivative of a variable. These occasionally got locked up in an infinite loop, where the event would fire over and over.