Fix poor test in test_Periodic.py
This test is checking whether we can detect local maxima using an event triggered off the derivative of that variable. A known difficulty is that, due to numerical integration errors, we the event may fire too often. Thus, the test should check whether it detects the maxima, whether it detects *only* the maxima.