Replace FORTRAN version of ddaskr integrator with C version, based on https://github.com/sjoelund/Cdaskr .
The trickiest part of doing so was fixing the definition of jac in ddaskr.pyf and the corresponding C function generated in Network_mod, so they matched the call expected by ddaskr. To be honest, I'm not sure why the previous version was working well with the Fortran code. Perhaps f2py can detect argument names in that case?